refactor(typosquat): split typosquat.rs into a submodule directory (#69)
src/enrich/typosquat.rs was 1197 LOC, over the 1000-LOC soft cap in the
module size budget. Split it into src/enrich/typosquat/ mirroring the
enrich/maintainer/ split from the #57 series: mod.rs owns the
TyposquatFinding type, the embedded list constants, and the public
enrich entry points, with ecosystem.rs (the SupportedEcosystem enum),
canonical.rs (name normalization), matching.rs (Jaro-Winkler and Maven
Levenshtein scoring), lists.rs (reference-list loading), and tests.rs.

Pure mechanical move, no behavior change. The public surface stays
reachable at crate::enrich::typosquat::* (TyposquatFinding, enrich,
enrich_with_threshold, SIMILARITY_THRESHOLD) and pub mod typosquat; in
enrich/mod.rs is unchanged. Cross-module helpers are pub(super).

Closes #69.
